// src/app/core/models/cda/inspectionDetailsBCSPCase.ts
export class InspectionDetailsBCSPCase {
    inspectionDetailsBCSPCaseID: number;
    aggreementIsChecked: boolean = true;
    applicationFileMasterID: number;
    investigationDate: string | null;
    landOwnerName: string;
    isLandUnderLease: boolean | null;
    leaseYear: number | null;
    actualProposedSizeNorth: string;
    actualProposedSizeSouth: string;
    actualProposedSizeEast: string;
    actualProposedSizeWest: string;
    actualProposedSizeLandQty: string;
    isApplicantLandOwner: boolean | null;
    landTopShil: string;
    rsKhotian: string;
    layOutPlotNo: string;
    landArea: string;
    proposedBuldingEnvelop1stFloor: string;
    proposedBuldingEnvelopTillTopFloor: string;
    proposedBuldingEnvelopFrom2ndFloor: string;
    proposedBuldingEnvelopStair: string;
    proposedBuldingEnvelopStairTotal: string;
    proposedBuldingEnvelopParkingSpace: string;
    isAnyExistingBuilding: boolean | null;
    existingBuildingNature: BuildingType | null;
    existingBuildingSize: string;
    structureToProsture: string;
    isNearbyOpenSpaceOk: boolean | null;
    connectedRoadName: string;
    connectedRoadWidth: string;
    connectedRoadDirection: string;
    connectedRoadType: string;
    connectedRoadProposedType: string;
    isLandSiteWorkStart: boolean | null;
    hasPreviousStructureApproval: string;
    has3BOrOtherNotice: string;
    isApplicationInformationOK: boolean | null;
    connectedMainRoadName: string;
    isInKpiArea30M: string;
    isHillLand: boolean | null;
    isHillNear: boolean | null;
    hIllDistance: string;
    proposedLandType: LandType | null;
    isVisibleAllStructureIn12M: boolean | null;
    visibleAllStructureIn12MNorth: string;
    visibleAllStructureIn12MSouth: string;
    visibleAllStructureIn12MEast: string;
    visibleAllStructureIn12MWest: string;
    isNearDrain: boolean | null;
    drainLength: string;
    drainType: DrainType;
    isEffectedNaturalWaterFlow: boolean | null;
    isPossibleToConnectDrain: boolean | null;
    hasElectricStructureIn10M: boolean | null;
    electricStructureIn10MDistance: string;
    nearElectricLineType: ElectricLineType | null;
    hasNearDefenceStructure: boolean | null;
    nearDefenceStructureDetails: string;
    isLakeInProposedLand: boolean | null;
    isHistoricalStructureInProposedLand: boolean | null;
    isLandFillFromLake: boolean | null;
    hasNaturalEffection: boolean | null;
    landClearanceNo: string;
    landSpecialClearanceNo: string;
    investigationOfficerID: number | null;
    investigationOfficerSignetureDate: string | null;
    deptHeadID: number | null;
    deptHeadSignetureDate: string;
    isDelete: boolean | null;
    tR_StructureUse: string;
    tR_StructureFloor: string;
    tR_StructureArea: string;
    tR_FloorSize1st: string;
    tR_FloorSizeTillTypical: string;
    tR_MGCGroundFloor: string;
    tR_MGCLimit: string;
    tR_LandClearanceNo: string;
    tR_LandClearanceReceiveDate: string | null;
    tR_LandClearanceSubmitDate: string | null;
    tR_LandSpecialClearanceNo: string;
    tR_LandSpecialClearanceReceiveDate: string | null;
    tR_LandSpecialClearanceSubmitDate: string | null;
    tR_ConnectedRoadName: string;
    tR_ConnectedRoadWidth: string;
    tR_ConnectedRoadWidthProposed: string;
    tR_RoadConcessions: string;
    tR_ThanaID: number | null;
    tR_MouzaID: number | null;
    tR_DesignCompany: string;
    tR_DesignCompanyMobile: string;
    tR_DesignEngineer: string;
    tR_DesignEngineerMobile: string;
    tR_VOIDNo: string;
    tR_VOIDSize: string;
    tR_FlatCount: number | null;
    tR_CarParkingCount: number | null;
    tR_BikeParkingCount: number | null;
    tR_AuthorisedBy: number | null;
    tR_AuthorisedDate: string | null;
    createdBy: number;
    createdDate: string;
    updatedBy: number;
    updatedDate: string;
    investigationOfficerName: string;
    investigationOfficerMobileNo: string;
    investigationOfficerSignature: string;
    deptHeadName: string;
    deptHeadMobileNo: string;
    deptHeadSignature: string;
    nearElectricLineTypeStr: boolean;
    drainTypeStr: string;
    proposedLandTypeStr: string;
    existingBuildingNatureStr: string | null;

    refNo: string | null;
    applicantName: string | null;
    rsNo: string | null;
    bsNo: string | null;
    thanaNameBangla: string | null;
    mouzaNameBangla: string | null;
}